Added Auth Functionality

+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -15,23 +19,28 @@ export class UsersService {
   // login(email: string, password: string) {}
 
   async findOne(id: number) {
+    if (!id) {
+      throw new BadRequestException('Invalid id.');
+    }
     const user = await this.repo.findOneBy({ id });
     if (!user) {
-      throw new NotFoundException('No users found matching this email.');
+      throw new NotFoundException('No users found with this id.');
     }
     return user;
   }
 
   async findAllByEmail(email: string) {
     const users = await this.repo.findBy({ email });
-    if (!users || !users.length) {
-      throw new NotFoundException('No users found matching this email.');
-    }
+    // if (!users || !users.length) {
+    //   throw new NotFoundException('No users found matching this email.');
+    // }
     return users;
   }
 
   async update(id: number, attrs: Partial<User>) {
-    console.log({ attrs });
+    if (!id) {
+      throw new BadRequestException('Invalid id.');
+    }
     const user = await this.findOne(id);
     if (!user) {
       throw new NotFoundException('User not found');
@@ -41,6 +50,9 @@ export class UsersService {
   }
 
   async remove(id: number) {
+    if (!id) {
+      throw new BadRequestException('Invalid id.');
+    }
     const user = await this.findOne(id);
     return await this.repo.remove(user);
   }
